Add config file flag to use external DB for timing data
This commit adds a new config flag to the stestr config file to specify
an external subunit2sql DB to use for historical scheduling data instead
of the data in the repository. There are some cases where there is a
subunit2sql repository but runners don't have write permission necessary
to use a full sql repository type. This commit enables users to query a
db for timing info instead of relying what's in the configured
repository.